refactor(db): rename database package to db
