+++ src/packages/db/.gitignore
@@ -6,6 +6,7 @@
 .pnp.js
 
 # database
+/supabase
 /prisma/db.sqlite
 /prisma/db.sqlite-journal
 db.sqlite
