+++ src/packages/db/package.json
@@ -20,6 +20,7 @@
     "@prisma/client": "^5.14.0"
   },
   "devDependencies": {
-    "prisma": "^5.14.0"
+    "prisma": "^5.14.0",
+    "ts-node": "^10.9.2"
   }
 }
