+++ src/packages/db/prisma/schema.prisma
@@ -1,0 +1,137 @@
+// This is your Prisma schema file,
+// learn more about it in the docs: https://pris.ly/d/prisma-schema
+
+generator client {
+  provider = "prisma-client-js"
+}
+
+datasource db {
+  provider  = "postgresql"
+  // NOTE: When using mysql or sqlserver, uncomment the @db.Text annotations in model Account below
+  // Further reading:
+  // https://next-auth.js.org/adapters/prisma#create-the-prisma-schema
+  // https://www.prisma.io/docs/reference/api-reference/prisma-schema-reference#string
+  url       = env("DATABASE_URL")
+  directUrl = env("DIRECT_URL")
+}
+
+model Account {
+  id                       String  @id @default(cuid())
+  userId                   String
+  type                     String
+  provider                 String
+  providerAccountId        String
+  refresh_token            String? // @db.Text
+  access_token             String? // @db.Text
+  expires_at               Int?
+  token_type               String?
+  scope                    String?
+  id_token                 String? // @db.Text
+  session_state            String?
+  user                     User    @relation(fields: [userId], references: [id], onDelete: Cascade)
+  refresh_token_expires_in Int?
+
+  @@unique([provider, providerAccountId])
+}
+
+model Session {
+  id           String   @id @default(cuid())
+  sessionToken String   @unique
+  userId       String
+  expires      DateTime
+  user         User     @relation(fields: [userId], references: [id], onDelete: Cascade)
+}
+
+model VerificationToken {
+  identifier String
+  token      String   @unique
+  expires    DateTime
+
+  @@unique([identifier, token])
+}
+
+model Country {
+  id             String                 @id
+  name           String                 @unique
+  latitude       Float
+  longitude      Float
+  stateProvinces StateProvince[]
+  users          User[]
+  alternatenames CountryAlternatename[]
+}
+
+model CountryAlternatename {
+  id          String  @id
+  countryId   String
+  name        String
+  isoLanguage String
+  isPreferred Boolean
+  isShortName Boolean
+  country     Country @relation(fields: [countryId], references: [id])
+}
+
+model StateProvince {
+  id             String                       @id
+  countryId      String
+  name           String
+  latitude       Float
+  longitude      Float
+  cities         City[]
+  country        Country                      @relation(fields: [countryId], references: [id], onDelete: Cascade)
+  users          User[]
+  alternatenames StateProvinceAlternatename[]
+}
+
+model StateProvinceAlternatename {
+  id              String        @id
+  stateProvinceId String
+  name            String
+  isoLanguage     String
+  isPreferred     Boolean
+  isShortName     Boolean
+  stateProvince   StateProvince @relation(fields: [stateProvinceId], references: [id])
+}
+
+model City {
+  id              String              @id
+  stateProvinceId String
+  name            String
+  latitude        Float
+  longitude       Float
+  stateProvince   StateProvince       @relation(fields: [stateProvinceId], references: [id], onDelete: Cascade)
+  users           User[]
+  alternatenames  CityAlternatename[]
+}
+
+model CityAlternatename {
+  id          String  @id
+  cityId      String
+  name        String
+  isoLanguage String
+  isPreferred Boolean
+  isShortName Boolean
+  city        City    @relation(fields: [cityId], references: [id])
+}
+
+enum Role {
+  CLIENT
+  ADMIN
+}
+
+model User {
+  id                  String        @id @default(cuid())
+  firstName           String
+  lastName            String
+  countryId           String
+  stateProvinceId     String
+  cityId              String?
+  emailAddress        String        @unique
+  password            String
+  role                Role
+  refreshTokenVersion Int @default(0)
+  accounts            Account[]
+  sessions            Session[]
+  country             Country       @relation(fields: [countryId], references: [id])
+  stateProvince       StateProvince @relation(fields: [stateProvinceId], references: [id])
+  city                City?         @relation(fields: [cityId], references: [id])
+}
