+++ src/packages/db/res/.gitignore
