+++ src/packages/db/scripts/generate-alternatenames.ts
